Refactor

+++ project/commons.cpp
@@ -22,7 +22,7 @@ int read_file_to_buffer(std::string &file_path, std::vector<char> &buffer) {
 
     std::ifstream file(file_path, std::ios::binary);
     file.seekg(0, std::ios::end);
-    int file_size = file.tellg();
+    long file_size = file.tellg();
     file.seekg(0, std::ios::beg);
     buffer.resize(file_size);
     file.read(&(buffer[0]), file_size);
