+++ project/evm_file_header.cpp
@@ -21,17 +21,21 @@ EvmHeader::EvmHeader() {
 void EvmHeader::read_header(std::vector<char> &buffer) {
 
     std::vector<char> header(buffer.begin(), buffer.begin() + header_length_in_bytes);
+
     std::vector<char> header_magic_number(header.begin(), header.begin() + magic_number_length);
     verify(header_magic_number);
     for (int i = 0; i < header_magic_number.size(); i++) {
         magic[i] = header_magic_number[i];
     }
+
     std::vector<char> header_code_size(header.begin() + magic_number_length,
                                        header.begin() + magic_number_length + header_param_length);
     code_size = extract_uint_32_value(header_code_size);
+
     std::vector<char> header_data_size(header.begin() + magic_number_length + header_param_length,
                                        header.begin() + magic_number_length + 2 * header_param_length);
     data_size = extract_uint_32_value(header_data_size);
+
     std::vector<char> header_initial_data_size(header.begin() + magic_number_length + 2 * header_param_length,
                                                header.begin() + magic_number_length + 3 * header_param_length);
     initial_data_size = extract_uint_32_value(header_initial_data_size);
@@ -50,7 +54,7 @@ uint32_t EvmHeader::extract_uint_32_value(std::vector<char> &buffer) {
     char *buffer_char = buffer.data();
     buffer_char[4] = '\0';
     uint32_t int_value;
-    int_value = (buffer_char[3] << 24) | (buffer_char[2] << 16) | (buffer_char[1] << 8) | buffer_char[0];
+    int_value = (uint32_t) (buffer_char[3] << 24) | (buffer_char[2] << 16) | (buffer_char[1] << 8) | buffer_char[0];
     return int_value;
 }
